refactor: separated search into different components
